fix: fix bugs after refactoring
- Fix Securely QR scanning bug
- Fix setBarFragment bug (could not open fragment without arguments)
- Fix lengthEdit bug (NullPointerException)
- Update Hex.java
- Update layout resource files

+++ app/src/main/java/com/merive/securely/utils/Hex.java
@@ -14,7 +14,7 @@ public class Hex {
     public static String encrypt(String value) {
         String[] encryptedSymbols = new String[value.length()];
         for (int i = 0; i < value.length(); i++)
-            encryptedSymbols[i] = Integer.toHexString((int) value.toCharArray()[i]);
+            encryptedSymbols[i] = Integer.toHexString(value.toCharArray()[i]);
         return String.join(" ", encryptedSymbols);
     }
